Bug fixes in generated code.

+++ Xbim.CodeGeneration/Templates/Infrastructure/ItemSetTemplate1.cs
@@ -45,7 +45,7 @@ namespace Xbim.CodeGeneration.Templates.Infrastructure
             #line hidden
             this.Write(@"<T> : Common.Collections.ItemSet<T>
     {
-        internal ItemSet(IPersistEntity entity, int capacity, byte property): base(entity, capacity, property)
+        internal ItemSet(IPersistEntity entity, int capacity, int property): base(entity, capacity, property)
         {
         }
